pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "macs-rna"
version = "0.1.0"
description = "Strand-specific MACS3 peak calling for RNA-seq data"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Ken Chen", email = "chenkenbio@gmail.com" },
]
dependencies = [
    "macs3>=3.0.0",
    "pyyaml",
]

[project.scripts]
macs-rna = "macs_rna.cli:main"

[project.urls]
Repository = "https://github.com/kenchenbio/macs-rna"
